# Try to initialize xmlXPathContext faster