**What problem is this PR intended to solve?**

A variation on solving the "slow xpath context creation" problem described in #3266. See first attempt at #3378.

Copying the struct and the hash tables is slightly faster than calling `xmlXpathNewContext`:

    Comparison:
        large: optimized:    17780.0 i/s
             large: main:    15143.4 i/s - same-ish: difference falls within error

    Comparison:
        small: optimized:    62891.9 i/s
             small: main:    49413.5 i/s - 1.27x  slower

But still slower than re-using the context object ("edge", v1.18.0.rc1):

    Comparison:
             large: edge:    22230.9 i/s
        large: optimized:    17780.0 i/s - 1.25x  slower
             large: main:    15143.4 i/s - 1.47x  slower

    Comparison:
             small: edge:   129162.0 i/s
        small: optimized:    62891.9 i/s - 2.05x  slower
             small: main:    49413.5 i/s - 2.61x  slower


I probably won't merge this, but thought it was worth sharing for posterity.

Next I'm going to try an extension of #3378 that will push and pop "eval frames" that include the internal state but also the registered functions/namespaces/lookup handlers.